Fixes + makefile

+++ source/mosaic/src/Mosaique2.hpp
@@ -7,4 +7,4 @@
 
 ImageBase get_corresponding_image(bool color, int m);
 ImageBase resizeImage(ImageBase& img, int newWidth, int newHeight);
-void mosaique2(ImageBase &imIn, ImageBase &imOut, int x0, int y0, float seuilVariance, int tailleMin);
+void mosaique2(ImageBase &imIn, ImageBase &imOut, int x0, int y0, int seuilVariance, int tailleMin);
